[chore] 부하 테스트 별도 패키지로 분리 (#33)
